requirements.txt
# requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
# pandas>=1.3.0 # Only needed if you add CSV saving or analysis
selenium>=4.0.0
webdriver-manager>=4.0.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0 # Recommend CPU version for easier install for users
# faiss-gpu>=1.7.0 # Optional: User can install manually if GPU desired/compatible
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.10.0 # Only needed if build_index uses it explicitly